requirements.txt
gTTs==2.22
Pillow==8.1.0
pyperclip==3 0.4.1
sounddevice==0.4.1 